detector/requirements.txt
kafka-python
python-arango